+++ src/Contracts/Customer.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Konekt\Customer\Contracts;
 
 use Illuminate\Database\Eloquent\Relations\MorphMany;
+use Konekt\Address\Contracts\Address;
 
 interface Customer
 {
@@ -23,4 +24,16 @@ interface Customer
     public function getName(): string;
 
     public function addresses(): MorphMany;
+
+    public function hasDefaultBillingAddress(): bool;
+
+    public function hasDefaultShippingAddress(): bool;
+
+    public function defaultBillingAddress(): ?Address;
+
+    public function defaultShippingAddress(): ?Address;
+
+    public function setDefaultShippingAddress(Address|int|null $address): void;
+
+    public function setDefaultBillingAddress(Address|int|null $address): void;
 }